Add support for setting root_route on stream; Bump to 0.2.5
`root_route` is a flag that says we should serve this stream up when a
request does not include stream information.  It can only point one place,
so setting it as a boolean on streams isn't ideal (the last one to set it
wins), but it will get the job done for now.