Slave: Send listeners one-at-a-time during handoffs
Attempt to work around a potential Node bug by only sending one listener at a
time, to lessen the chance that a listener could disconnect between the time
we go to send them and the time they are finally sent via IPC.

Node internally queues sends with handles to avoid OS issues, and doesn't do
enough checking before it goes to convert the socket once it is time to send,
so it's possible to feed in a valid socket and have it be invalid by the time
Node goes to handle it.  If we only queue one at a time and do our own check
right at send time, we should be ok.