Refine handle passing during slave handoffs
During a slave mode handoff, we have the tricky task of passing a handle from
a cluster worker to the cluster master, on to the new cluster master, and then
on to the new cluster workers.

As of Node 0.10, normal server handles don't seem to do well with that.

We're now passing the underlying handle from old -> new without the server
wrappings (so that the child_process code doesn't trn on listening in the new
cluster master), and then passing the fd number on to the cluster workers.