Fix handoff error if last listener was lost before sending
When a handoff starts, each slave puts together an array of listeners that
are "boarding"—waiting to be sent to the new process.  When it gets an ack
that the listener arrived, it runs a function to send the next listener out
of that list.  When there are no more listeners waiting to board, the ack
function triggers the callback.

If a listener has disconnected in the time between assembling the list and its
turn for flight, the sending code skips the listener and moves on to the next
one.

The bug was that if the last listener in the boarding line had dropped, the
ack function never got a chance to trigger the callback.  This fix adds an
additional spot to trigger the callback in that situation.