Master.Stream: Add `getStreamKey` function
In Standalone mode, the `Slave.Stream` object is given a `Master.Stream`
object as its source, rather than the `Slave.SocketSource` object that
it would get in a master/slave setting.  If preroll is enabled, the
preroll object checks `source.getStreamKey()`, which was defined on
`SocketSource` and on the actual source objects, but not on
`Master.Stream`. This fix adds the function to that class, fixing prerolls
in standalone mode.

(Thanks for the spot, @jcohenho!)