Add alerts for `slave_unsynced` and `slave_unresponsive`
Every five seconds we poll slave status. This code adds checks against the
master's rewindbuffer status, and alerts is the slave is not responding to
our status check or contains data that is out-of-sync with the master.

While currently it just alerts, in the future these checks could cause a
slave to be re-initialized or disconnected.