Pass custom UA information through HLS requests
If it's possible to set a custom UserAgent with the IOS AVPlayer, we haven't
found it, so we're using a ua query string to set a custom UA for our client.
Previously, that UA had only gone to the session start, which gets it correctly
set on the session, but doesn't help us in looking at listens.