Slave/Shoutcast: Hopefully clear up a memory leak
Once adding Shoutcast listeners to our production server yesterday we noticed
memory use jumping. Local testing shows this commit clearing up a circular
reference between the Shoutcast output and its Rewinder object.

It also adds the option to spawn the slave with
`----enable-webkit-devtools-slaveworker` and get the slaveworker listening
for inspection (you need to be running `cluster: 1` for this to work).