Update RPC library; Trace down a test issue
Update `ipc-rpc` to 0.1.3 to fix an issue where exceptions in a callback
(such as a failing mocha assertion) were getting silently swallowed and
causing mysterious timeouts.

Unfortunately, that just let me track down a different issue that was causing
tests that work when run independently to not pass when running all tests.
Apparently the global nature of Node's built-in `cluster` causes all Slave
Mode instances to try and use the same cluster pool, which means we are
getting responses from the wrong set of workers.

That's going to take a little more work to get around, but only affects tests.