DumpRestore: Don't allow dumps while RewindBuffer is loading
This is part of trying to make sure unexpected things don't happen during
handoffs. Dumping while we have a partial buffer would break our ability
to restore if we were to crash.

FIXME: What we really need is the ability to not dump when we've been
told to expect data via a handoff, but haven't gotten any yet. In that
case, `loadBuffer` hasn't been called, so `isLoading()` would still
be false.