Slave: Bubble connection count and kbytes sent to status
Add `connections` and `kbytes_sent` to status object for slave streams, and
expose a rolled-up number in a `_stats` object.

Slave WorkerPool now queries SlaveWorker status every second, with the intent
that the above data could be useful for listener distribution decisions.
Currently, that's still done via random sampling of the eligible pool.