Add API Port to standalone mode; Rebuild Standalone handoff
Start to make standalone mode a viable production option by allowing the API
to run on an alternate port. Allows the main port to be safely exposed to
the world while still keeping the API port behind a firewall.

Rebuild standalone handoff support, bringing in many of the RPC call changes
already implemented in master and slave modes. Rebuild support for sending
listeners to the new standalone process.

* Master: Fix a bug where handoff send would hang if a stream's buffer
  was empty.
* Base: Remove deprecated support for internal handoffs. Use the included
  runner command for handoffs instead.
* Mode.Master: Comment out an unused but broken RPC function. Clean up some
  callback naming confusion.