Add manual stream_key option to enable AAC preroll support
Historically, we've grabbed encoding information from MP3 and AAC headers,
and used that for things like making sure incoming sources match our
expectations, asking for preroll, etc. We call the short, stringified
version of this the stream key.

Unfortunately, there are important pieces of AAC encoding information that
aren't in the header, including bitrate and AAC-HE or AAC-HEv2.

As a stop-gap, allow a manual `stream_key` configuration option to set
the key that should be sent to preroller for a given stream.