Trigger preroll impression based on connected time, not delivered time
Instead of passing an impression callback to the client output and
relying on it to trigger based on the delivered duration, trigger an
impression if the client hasn't disconnected by a configured delay.

This allows the impression logic to be contained inside Preroller,
and simplifies the impression criteria by removing burst amount as
a factor.