Fix a bug sending prerolls to Shoutcast clients
In the Shoutcast output, the Icy injector wasn't getting connected until
we had a Rewinder, meaning that the pipe for our preroll data would get
stopped via stream backpressue. That caused the preroll session to time
out, but we would then inject the rest of the data later once the stream
started flowing.

Attach the icy injector to the stream right away to relieve that pressure.

* Preroller: Rewrite `pump` function as a class, hopefully making it a
  little bit clearer what's happening where.
* Master.Stream: Add support for recreating a failed transcoding source
* Sources.Transcoding: Add debugging output for domain errors